==> CDD_training_aids/New_Stuff/Blood/1_Load_review.Rmd <==
---
title: "Load & Review Data v2 (get all feet sorted)"
author: "Wesley"
date: "2022/03/28"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load all of the Blood Data

To start, we need to load all of the Excel files. To do this, we will use the
readxl package. In addition, we're going to need to manipulate strings in 
the following, so let's pre-emptively load that package.

```{r libraries, error = FALSE, message = FALSE, warning = FALSE}
library("readxl")
library("stringr")
```

### Start with Blood 1, as Template

Blood 1 has many sheets, each of which is a nicely organized listing of VOCs
as detected by the GCxGC system's built-in scanning software, and exported. 
We will load them all. 

```{r load_data}
sheets <- excel_sheets("/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood1.xlsx")  # note: spaces in names
Blood1 <- lapply(sheets, FUN = function(x) {
           read_xlsx(path = "/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood1.xlsx",
                     sheet = x) })
sheets <- str_replace_all(sheets, " ", "_")
names(Blood1) <- sheets
```

Having things organized by sample is a good thing, but we really will
find it annoying finding things long-term. In the following, I will take each
of the lists of compounds relevant to Blood 1, and merge them into a single
data.frame object, so we can then filter that frame down to get a specific
compound, or a specific day/observation, and so on.

Before we do this, the first sheets has the observations duplicated across
the columns (wups!). Removing that:

```{r remove_sheet1}
Blood1[[1]] <- Blood1[[1]][, 1:ncol(Blood1[[2]])]
names(Blood1[[1]]) <- names(Blood1[[2]])  # fix names of sheet 1
```

```{r merge_sheets}
Blood1m <- do.call("rbind", Blood1)
names(Blood1m) <- c("PeakNum", "Sample", "Name", "Formula", "RTs", "1stRT","2ndRT","Similarity",
                   "Area", "Height", "QuantMass", "BaseMass", "QuantSN", "PeakSN",
                   "Type", "Conc", "ConcUnits")
```

A quick print of the data structure shows that the RTs, QuantMass and BaseMass
won't be usable as numbers. The RTs need to be split, as well.

```{r split_RTs}
RT1 <- as.numeric(
         unlist(
           lapply(strsplit(Blood1m$RTs, ","), FUN = "[[", 1)
         )
       )
RT2 <- as.numeric(
         unlist(
           lapply(strsplit(Blood1m$RTs, ","), FUN = "[[", 2)
         )
       )
Blood1m <- data.frame(Blood1m[, c(1:4, 6:17)], RT1 = RT1, RT2 = RT2)
```

```{r samples}
unique(Blood1m$Sample)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood1.rda", Blood1m)
```

### Now Do Blood 2

```{r load_data2}
sheets <- excel_sheets("/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood2.xlsx")  # note: spaces in names
Blood2 <- lapply(sheets, FUN = function(x) {
           read_xlsx(path = "/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood2.xlsx",
                     sheet = x) })
sheets <- str_replace_all(sheets, " ", "_")
names(Blood2) <- sheets

Blood2m <- do.call("rbind", Blood2)
names(Blood2m) <- c("PeakNum", "Sample", "Name", "Formula", "RTs","1stRT","2ndRT", "Similarity",
                   "Area", "Height", "QuantMass", "BaseMass", "QuantSN", "PeakSN",
                   "Type", "Conc", "ConcUnits")
```

```{r split_RTs2}
RT1 <- as.numeric(
         unlist(
           lapply(strsplit(Blood2m$RTs, ","), FUN = "[[", 1)
         )
       )
RT2 <- as.numeric(
         unlist(
           lapply(strsplit(Blood2m$RTs, ","), FUN = "[[", 2)
         )
       )
Blood2m <- data.frame(Blood2m[, c(1:4, 6:17)], RT1 = RT1, RT2 = RT2)
```

```{r samples2}
unique(Blood2m$Sample)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood2.rda", Blood2m)
```

### Now Do Blood 3

```{r load_data3}
sheets <- excel_sheets("/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood3.xlsx")  # note: spaces in names
Blood3 <- lapply(sheets, FUN = function(x) {
           read_xlsx(path = "/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood3.xlsx",
                     sheet = x) })
sheets <- str_replace_all(sheets, " ", "_")
names(Blood3) <- sheets

Blood3m <- do.call("rbind", Blood3)
names(Blood3m) <- c("PeakNum", "Sample", "Name", "Formula", "RTs","1stRT","2ndRT", "Similarity",
                   "Area", "Height", "QuantMass", "BaseMass", "QuantSN", "PeakSN",
                   "Type", "Conc", "ConcUnits")
```

```{r split_RTs3}
RT1 <- as.numeric(
         unlist(
           lapply(strsplit(Blood3m$RTs, ","), FUN = "[[", 1)
         )
       )
RT2 <- as.numeric(
         unlist(
           lapply(strsplit(Blood3m$RTs, ","), FUN = "[[", 2)
         )
       )
Blood3m <- data.frame(Blood3m[, c(1:4, 6:17)], RT1 = RT1, RT2 = RT2)
```

```{r samples2}
unique(Blood3m$Sample)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood3.rda", Blood3m)
```
