==> CDD_training_aids/New_Stuff/Blood/2_Normalize.Rmd <==
---
title: "Normalize Data"
author: "Wesley Burr"
date: "29/03/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
```

## Normalize Data

We will load each of the data files, identify the 'primary' peak for
Bromobenze, then normalize all Areas to be relative areas to the
reference standard.

```{r}
load("/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood1.rda")
bromo <- Blood1m %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(Area == max(Area))
split_Blood <- Blood1m %>% 
                group_by(Sample) %>%
                group_split()
for(j in 1:length(split_Blood)) {
  split_Blood[[j]]$Area <- split_Blood[[j]]$Area / unlist(bromo[j, "Area"])
}
Blood1m <- do.call("rbind", split_Blood)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood1n.rda", Blood1m)
```

This worked, now let's just loop the others to save writing:

```{r}
for(k in 2:3) {
  load(paste0("/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood", k, ".rda"))
  Blood <- get(x = paste0("Blood", k, "m"))
  bromo <- Blood %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(Area == max(Area))
  split_Blood <- Blood %>% 
                group_by(Sample) %>%
                group_split()
  for(j in 1:length(split_Blood)) {
    split_Blood[[j]]$Area <- split_Blood[[j]]$Area / unlist(bromo[j, "Area"])
  }
  Blood <- do.call("rbind", split_Blood)
  assign(x = paste0("Blood", k, "m"), value = Blood)
  save(file = paste0("/home/wburr/Rushali_Analysis/New_Stuff/Blood/Blood", k, "n.rda"), list = paste0("Blood", k, "m"))
}
```
